BST practice

+++ app/src/main/java/com/karnamic/dsainkotlin/trees/TreeNode.kt
@@ -1,7 +1,7 @@
 package com.karnamic.dsainkotlin.trees
 
 data class TreeNode(
-    var `val`: Int,
+    var `val`: Long,
     var left: TreeNode? = null,
     var right: TreeNode? = null
 )
