+++ app/src/main/java/com/karnamic/dsainkotlin/trees/binary_tree/LevelOrder.kt
@@ -11,15 +11,15 @@ import com.karnamic.dsainkotlin.trees.TreeNode
  * Output: [[3],[9,20],[15,7]]
  */
 
-fun levelOrder(root: TreeNode?): List<List<Int>> {
-    val res = mutableListOf<List<Int>>()
+fun levelOrder(root: TreeNode?): List<List<Long>> {
+    val res = mutableListOf<List<Long>>()
     if (root == null) return res
 
     val q = ArrayDeque<TreeNode>()
     q.add(root)
 
     while (q.isNotEmpty()){
-        val level = mutableListOf<Int>()
+        val level = mutableListOf<Long>()
         val qLength = q.size
 
         for(i in 0 until qLength) {
